Add double suppression size and noise (optional)
Adding optional noise is necessary due to the switch from float based probabilit
maps to int. The switch means that the probability map has only 258
possible values, increasing the number of maxima per nms window
significantly as multiple maxima have the same value.
This changes the behaviour drastically, especially if the classifier
is of lower quality. The second change gives the possibility
to change the window size of the second nms pass which is
particularly useful for datasets of different resolution and
multiple double detections per window (noisy classifiers).