Pass databse client to write_done and check func
Not passing the client and establishes a new connection to the db
everytime its called. This leads to problems when solving with many
workers as we open and close sockets too fast such that we run out of
ports eventually. This happened sometimes for the write_done call which
leads to blocks being solved twice. This is a problem because we remove
all constrains from fully solved triplets during solving (w.o. this bug
these can only exist in the context area and are thus not written).
Solving twice could thus cause violation of constraints. This commit
fixes the primary error but it would be even better to make edge and
vertex updates atomic with the block_done call.